updateLogs

+++ frontend/src/pages/SignUpPage.js
@@ -6,34 +6,39 @@ import Button from 'react-bootstrap/Button';
 import { Helmet } from 'react-helmet-async';
 import { useContext, useEffect, useState } from 'react';
 import { Storage } from '../Storage';
-import { toast } from 'react-toastify';
-import { getError } from '../UtilityE';
 
-export default function LoginPage() {
+export default function SignUpPage() {
   const navigate = useNavigate();
   const { search } = useLocation();
   const redirectInUrl = new URLSearchParams(search).get('redirect');
   const redirect = redirectInUrl ? redirectInUrl : '/';
 
+  const [name, setName] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
 
   const { state, dispatch: ctxDispatch } = useContext(Storage);
   const { userInfo } = state;
 
   const submitHandler = async (e) => {
     e.preventDefault();
+    if (password !== confirmPassword) {
+      console.log('Invalid Entry');
+      return;
+    }
+
     try {
-      const { data } = await Axios.post('/api/users/signin', {
+      const { data } = await Axios.post('/api/users/signup', {
+        name,
         email,
         password,
       });
-
       ctxDispatch({ type: 'USER_SIGNIN', payload: data });
       localStorage.setItem('userInfo', JSON.stringify(data));
       navigate(redirect || '/');
     } catch (err) {
-      toast.error(getError(err));
+      alert('Invalid Entry');
     }
   };
 
@@ -45,10 +50,15 @@ export default function LoginPage() {
   return (
     <Container className="s-container">
       <Helmet>
-        <title>Log In</title>
+        <title>Make New Account</title>
       </Helmet>
-      <h1 className="my-3">Account</h1>
+      <h1 className="my-3">Make New Account</h1>
       <Form onSubmit={submitHandler}>
+        <Form.Group className="mb-3" logid="name">
+          <Form.Label>Name</Form.Label>
+          <Form.Control onChange={(e) => setName(e.target.value)} required />
+        </Form.Group>
+
         <Form.Group className="mb-3" logid="email">
           <Form.Label>Email</Form.Label>
           <Form.Control
@@ -66,12 +76,21 @@ export default function LoginPage() {
             onChange={(e) => setPassword(e.target.value)}
           />
         </Form.Group>
+        <Form.Group className="mb-3" logid="confrimPassword">
+          <Form.Label>Confirm Password</Form.Label>
+          <Form.Control
+            type="password"
+            onChange={(e) => setConfirmPassword(e.target.value)}
+            required
+          />{' '}
+        </Form.Group>
+
         <div className="mb-3">
-          <Button type="submit">Log In</Button>
+          <Button type="submit">Sign Up</Button>
         </div>
         <div className="mb-3">
-          First time?{' '}
-          <Link to={`/signup?redirect=${redirect}`}>Create Account</Link>
+          Returning User?{' '}
+          <Link to={`/signup?redirect=${redirect}`}>Log In</Link>
         </div>
       </Form>
     </Container>
